+++ src/egdata.h
@@ -152,7 +152,7 @@ extern int16 g_pitchMatrix[9];
 extern int16 g_rollMatrix[9];
 extern int16 g_matrixScratch[];
 extern int16 g_ourHead;
-extern int g_ourPitch;
+extern int16 g_ourPitch;
 extern int16 g_ourRoll;
 /* g_viewZ: altitude-Z. int16 for most TUs; egtarget.c reads the same storage
    as uint16 (it #defines G_VIEWZ_UNSIGNED before including this header). */
